Backend for the Rating , Changing color according to rating

+++ app/src/main/java/model/DishReview.kt
@@ -1,6 +1,5 @@
 package model
 
-import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.GeoPoint
 import com.google.firebase.firestore.ServerTimestamp
 import java.util.Date
@@ -9,7 +8,7 @@ data class DishReview(
     var id: String = "",
     var dishName: String = "",
     var price: Int = 0,
-    var rating: Double = 0.0,
+    var rating: Int = 0,
     var likedBy: List<String> = emptyList(),
     var likes: Int = 0,
     val imageUrl: String = "",
